+++ src/main/java/maroonshaded/gildedarmor/item/GildedArmorItem.java
@@ -1,17 +1,10 @@
 package maroonshaded.gildedarmor.item;
 
-import net.minecraft.ChatFormatting;
-import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ArmorItem;
 import net.minecraft.world.item.ArmorMaterial;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.level.Level;
-
-import javax.annotation.Nullable;
-import java.util.List;
 
 public class GildedArmorItem extends ArmorItem
 {
@@ -24,12 +17,5 @@ public class GildedArmorItem extends ArmorItem
     public boolean makesPiglinsNeutral(ItemStack stack, LivingEntity wearer)
     {
         return true;
-    }
-
-    @Override
-    public void appendHoverText(ItemStack pStack, @Nullable Level pLevel, List<Component> pTooltipComponents, TooltipFlag pIsAdvanced)
-    {
-        for (int i = 1; i <= 3; i++)
-            pTooltipComponents.add(Component.translatable("item.gildedarmor.generic.upgrading_note" + i).withStyle(ChatFormatting.GRAY));
     }
 }